Implemented dyanmic queue counter
TODO: fix db bug
